Add seed file for dummy client database
This current only seeds the events as there were already users in the table
but I will add seeds for other tables as I come across the need for them.
There will eventually be seeds for all tables that will fit the specs so we can
clear and repopulate the dummy client database as and when we need.