WIP saving target database credentials
Aside from the demo requirements, some of the constraints won't be an issue when it is self hosted:
When starting up for the first time when the user registers, it will not try and connect to the target database until they have registered the target database (they should not be able to get to the dashboard until a successful connection has been made)

I _think_ the issues you were encountering with the inconsistency of the target database were because the credentials weren't saving correctly? The fix I have put in is very ugly and needs some tidying up/making prettier but the general logic/concept is there. I don't think Devise liked having the first_name and last_name fields saved differently from the target_database_* fields.

Let me know if this change makes any difference to you. Currently, as we are saving the credentials with the SensitiveData module, the target database is not saved against a user so if you want to switch database type (i.e. from mysql to postgres) then you need to re-register with the new database. This functionality is frustrating for now but is the correct functionality for the final application. As it will be self hosted, there should not be a need to change databases regularly. The ability to have multiple database dashboards will be in version two or three :D